export types to local file
