+++ src/orders/components/OrderCustomerAddressesEditDialog/types.ts
@@ -1,0 +1,15 @@
+import { AddressInput, AddressTypeEnum } from "@saleor/types/globalTypes";
+
+export interface OrderCustomerSearchAddressState {
+  open: boolean;
+  type: AddressTypeEnum;
+}
+export interface OrderCustomerAddressesEditDialogOutput {
+  shippingAddress: AddressInput;
+  billingAddress: AddressInput;
+}
+export enum AddressEditDialogVariant {
+  CHANGE_CUSTOMER,
+  CHANGE_SHIPPING_ADDRESS,
+  CHANGE_BILLING_ADDRESS
+}
